feat: add cast, crew, and orchestra pages with progress indicators and styling updates
