+++ src/_styles/base.pcss
@@ -57,3 +57,23 @@ ol {
 ul {
 	list-style-type: disc;
 }
+
+progress[value] {
+	/* Reset the default appearance */
+	-webkit-appearance: none;
+	appearance: none;
+
+	width: 100%;
+	height: 0.25em;
+
+	&::-webkit-progress-bar {
+		background-color: var(--c-bg-3);
+		border-radius: 2em;
+	}
+
+	&::-webkit-progress-value {
+		background-color: var(--c-fg-3);
+		border-radius: 2em;
+		/* transition: width 1s ease; */
+	}
+}
